pyproject.toml
[tool.poetry]
name = "qstash-python"
version = "2.0.0"
description = "Python SDK for Upstash QStash"
license = "MIT"
authors = ["Upstash <support@upstash.com>"]
maintainers = ["Upstash <support@upstash.com>"]
readme = "README.md"
repository = "https://github.com/upstash/qstash-python"
keywords = ["QStash", "Upstash QStash", "Serverless Queue"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Database",
    "Topic :: Database :: Front-Ends",
    "Topic :: Software Development :: Libraries",
]

packages = [{ include = "upstash_qstash" }]

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.23.0, <1"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
python-dotenv = "^1.0.1"
pytest-asyncio = "^0.23.7"
mypy = "^1.10.0"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
